pyproject.toml
[tool.poetry]
name = "core-ds-transaction-scoring"
version = "0.1.0"
description = ""
authors = ["Evgeny Grigorenko <egrigorenko@ditio.no>"]

[tool.poetry.dependencies]
python = "~3.12"
pydantic = "^2.5.2"
pymongo = {extras = ["srv"], version = "^4.6.1"}
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
